mir: Add a rewrite pass to fuse filter nodes
Generalize out a `fuse_nodes` function from the `fuse_project_nodes`
rewrite pass, and use this function to implement a new
`fuse_filter_nodes` rewrite pass which does the same thing, except it
fuses filter nodes by combining the conditions using an AND. This ought
to improve execution time for queries with multiple successive filter
nodes in the same way as for project nodes.
